+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- LLM evaluation logic
+
 ## [0.3.3] - 30.10.2024
 
 ### Changed
@@ -38,6 +41,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Swarm copy POC.
+- Agent memory.
 
 
 ## [0.2.0] - 22.10.2024